Fix list::apply, which broke when methods were removed.
This time, add a test in sanity.ly to make sure this problem doesn't happen
again.

Also, make sure to prepare the registers of the callee function, instead of
hoping that they're right. That was never done, so it was really a miracle
that list::apply ever seemed to work.