Make show less painful and awful. Some comments updated.
This was done to prepare it to be called from a function. Also, o_new_instance
and o_get_function now work properly in show.

I decided to use a giant table to hold all opcode information. The opcode is
at the front of each entry so that it can easily be sanity-checked by looping
over the opcode table and checking opcode_table[i][0] == i.

I tried splitting show off into functions, only to be horrified at the mess
that resulted and backtracking on it. So I'll just make show into a function.