Show now nicely dumps instance values. Fixes #56.
It now prints the name of the property (as well as where it inherited from),
but doesn't print a tag if it's just the current class.

For a Point class with no parent, the properties are simply printed.
For a ValueError class (which has Exception as a parent), properties are
listed as being inherited from Exception.

This code has a bug, however: If a class has properties and the parent does,
all properties will be printed under the parent. I intend to leave this in
until classes are more stable (so I don't have to redo the fix if properties
change or somesuch).

Oh, and user-defined classes are printed backwards. Whoops!