Classes: Functions defined inside the ::new of a class now linked into it.
Previously, this:

class Point(integer x, integer y) { function f() { } }

would not result in 'f' being made available only within the Point class. Now,
f is pulled into the functions provided by Point. However, there are things
about it that don't work (A class instance is not the first value, for
example).

It's a step in the right direction though. :)