Classes: Functions have implicit self and class names for trace.
A function defined within a class is now assumed to have self as the first
var. This is currently not accessible (the name is not a valid var name), but
that's so that self can be reached via @ in the future.

Additionally, native function vars are now created with a class name if there
is a class name. Now, exceptions that are raised inside of a class function
will print the class name and their own name, making it clear that they're
part of a class.

I've also made it so that blocks now contain 'self'. The intent is to use this
to make nested classes possible (because if nested functions are possible,
then nested classes should be as well).