Classes: Property creation, get, and set working. Fragile!
class Point(integer xx, integer yy) { integer @x = xx, @y = yy }

Now, the interpreter creates @x as property 'x' within the Point class.
Within the constructor to Point, the property x is only available as @x. This
is done so it's obvious that a property of a class is being used and not some
global or local variable somewhere. The idea is borrowed from Ruby.

Outside of the class, @x is accessible through instances of Point using the
'<value>.x' notation. The interpreter also checks that a class doesn't make
multiple properties of the same name. Additionally, the interpreter requires
a value when creating a new property.

This is not an accident. In the future, I intend to have all assignments
require a value, (but only after an Option class and a None value have been
provided).

Classes are still fragile, but at least they're getting better.