Classes: Values passed to the constructor are no longer properties.
I actually don't care for this too much. So remove it, making it where only
properties declared -within- the constructor become properties. I ended up not
liking automatically making things properties, because now every property has
to have a @ prefix to be used.

Also, user-defined class properties are now linked from first to last (with
the properties of a class holding the start). A class now holds a prop_count,
the number of properties it has instead of a start (the start wasn't used
anyway).