Classes: Calls to member functions automatically get self.
This was fairly simple to do: A function call already adds the function as the
first argument (and then later picks it off). This simply checks if the tree
holding the function should get 'self' instead (and count it as an argument).

This does a basic test to prevent calls to global functions from getting self.
This test will need to be improved when user-defined classes can nest as well
as inherit from each other. Consider:

Class A() {
    function f() {   }
    class B() inherits A {
        f() # This should work, since B inherits A and A defines f.
    }
}

Hurray for classes starting to be less fragile!