Parser: Function capture now recurses to capture. Fun! Fixes #57.
So I had a fun bug in classes:

'''
class Point() { function f() {} function g() { f() } }
Point p = Point::new()
p.g() # Error: Got 1 arg, expected 0.
'''

Huh? Both f and g are defined within a class, and should both get an implicit
'self'. Inspecting further showed that...

class Point entered...
    function f processed
class Point exited...
    function g processed

The problem is that function collection was simply continuing over the same
loop. When function f exited, the } that closed it was yielded up to the
caller. This is why Point thought it was closed when it wasn't.

Because of this, recursion is now used for handling function collection. I've
also tested the same example in lily_cliexec before committing this to make
sure this works (also, the implicit self passing seems okay too).

With this, I consider user-defined classes to be stable enough to push to
origin and close #57.