Merge o_build_list and o_build_tuple into one opcode.
While writing o_new_instance for classes, I realized that these two could be
merged by making the initialization loop use rhs->sig since that will always
be the proper thing: For lists, it'll never change, for tuples it'll always be
what that particular index should be.

Hurray for a simpler vm!