Register info no longer holds class info. Show properly dumps funcs now.
I don't think the former has really been working for a while. I'm pretty sure
it would have stopped working when functions got their own table in the vm's
function table. Anyway, even if it DID work, there's a better way to get the
class name of a function given.

The better way is to print the class info of a particular function when within
'show_function' of lily_debug. This has access to functions called directly,
and can print class info on them. I've tested this with list::append to make
sure it works.

This also saves a little memory, as functions no longer have a spot in
register info for holding class information.