Implement 'calltrace(=>list[tuple[string, integer]])'. Fixes #52.
I decided to put this alongside show, because why not? Additionally, the bug
mentions that the result should be a list of 'tuple[string, string, integer]'
(two strings). The reason that this wasn't done is because there is no import
(and thus no filename to print). I'll update calltrace when import works.

This is a handy function for determining the current Lily call stack. This
could later be used for debugging (ex: A function that shows the call stack
before calling a class function to determine what is calling that class
function).

Note: Built-in functions have the integer part as -1, which I might fix to 0
later. Lily is different in that it adds foreign function entries to the call
stack. I may also make it where calltrace doesn't add itself to the trace
later too.