Generics: First attempt. Some stuff works, other stuff is fragile.
What works:
function myappend[A](list[A] lst, A value => list[A]) {
    lst.append(value)
    return lst
}

The interpreter is able to call this function with [1] and 1 or [1.1] and 2.2
and it works just fine. However, there's various things that don't work with
generics:

The interpreter is able to handle generic function parameters, but it can't
handle vars or temporaries inside of a function that use generics.

Generics always start with A and go up in sequentially alphabetically. This
may be changed to allow names for generics in the future. I sort of doubt it
thought: I like the consistency of forcing A, B, C...

Some hacks were done in symtab to keep scan_seed_sig from damaging template
parameters (part of the problem there is that scan_seed_sig can be removed
after generics are stable, but until then it needs to be bandaged).

The symtab fools the parser by switching the sig of the template class when
the parser asks for a generic class by name. It's rather nifty, I think.