Generics: Local vars and storages are now resolved in generic functions.
This is a really, really big chunk of getting generics working. Before this,
only the parameters of a generic function would get the proper types. The
local vars and storages of a generic function would be left as-is (and likely
crash/behave strangely).

function f[A](list[A] lst) { tuple[list[A], list[A]] t = <[lst, lst]> }

will now work. If called with a list[integer], then t will be
tuple[list[integer], list[integer]].

With this done, the rest of generics is getting different parts of emitter to
stop doing sig==sig (and template matching), and adjusting the generics
available when a function enters/exits.