Generics: Fix template resolution in eval, call returns, oops in vm resolve.
The first one was that calls were checking only for template_pos of a sig
being non-zero instead of that OR the template class (because A is a template
sig with a template pos of 0.

The second was an issue where call returning was grabbing the wrong signature.
It wrongly assumed that it could just pull from the call signature instead of
properly using the emitter's sig stack.

The final issue was not updating the register count after doing a generic
function resolution.