Classes: First pass at combining generic constructors and classes.
Creating a generic class is pretty simple:
    class Point[A](A value) { A @value = value }

With this patch, the interpreter now recognizes that Point contains other
elements, and that 'Point[integer]' is a valid class identifier. Point::new
in this case works, but the constructed instance is very confused about the
types it has (because they're all wrong).

Generic classes still don't technically work, but this is a step toward the
right direction.