Un-break sys::argv, add tests for sys::argv usage.
The vm was starting the prep from __main__ if there had been no prep
previously. This was fine until the symtab's vars became a 'most recent first'
kind of linked list (which meant that sys wasn't loaded and using it would
cause a crash).

The code for handling multiple packages in parser is gone. I'm not going to
do anything more to packages just yet, because I'm not sure -exactly- how I
want to do packages. I'm waiting for 'import', specifically.