Emitter: Remove impossible check within eval_call, GLOBAL_LOAD_CHECK.
It turns out that it's simply impossible (I think) to have a function call with
the call being a global var. This is because functions have a depth of -1 so
they get marked as readonly (and thus loaded from the vm's func table).

I tried this using 'function f(){}    var v = f    v()' and still could not
trigger this check. I seriously don't think it's possible.

This removes the last user of GLOBAL_LOAD_CHECK, so get rid of that too.