Emitter: For loop iteration no longer targets global vars (Fixes #65)
Previously, for loops all assumed that the var given was a local. If the loop
was within __main__, this just resulted in the for loop treating the global var
as a local. Within __main__, this is kind of harmless.

Within a non-__main__ function, the global var's register spot was interpreted
as the spot of a local register. This could result, in, say, the for loop
thinking that the counter was also the loop start (peculiar). In the case of
a global var with a high register spot...I'm assuming there would be a crash.

Anyway, the fix is to create a storage to do the iteration, and do an
o_set_global after the prep and each loop pass. Easy.