Fix "list[integer] l = [1].append(10)" crash.
This was caused by a bug when append tried to swallow the [1] of the binary
right side as an argument. The append tree was not updating its parent.

Since append did not see that it had a parent, the assignment would use a
null result (causing a crash). A fix was given to oo assign too, to keep it
from crashing on compound ops.

A test has been added to pre-commit-hook.py to keep this from occuring again.