Make sure nested classes actually, you know, work. Test added!
    class c1(integer v1) {
        class c2(integer v2) {
        }
        var @v = c2::new(1)
        }
    }
    var v = c1::new(1)

This snippet was reporting that c1::new takes 2 arguments, not one. So the first
thing that was done was to make sure that a nested class constructor doesn't
automatically get a 'self' of the parent class...because that's wrong.

But then calling c2::new started to crash. It turns out that the reason was that
the vm was not loading the functions of scoped out classes. This was trivial to
fix.

Added a test to pre-commit-hook.py to prevent this from happening again.