Rip function decl and regular decl parsing apart.
So...since the beginning of Lily, the proper way to enter a function has been
to call collect_var_sig with CV_TOPLEVEL and CV_MAKE_VARS. Later it was
changed to take a starting class too.

When classes came along, that was added to: To enter a function, call
collect_var_sig to collect the function type (and at the same time enter the
function but don't mention that).

That functionality has now been stripped from collect_var_sig. The proper way
to enter a function (internally) is to call...parse_function. This parses the
header and actually enters the function. No more, no less.

I've also simplified the logic for handling generics and stuff for functions.
On the outside, there is no change. Internally, a function that did two things
has been cut into two functions.

Use collect_var_sig to collect sigs only now.

I did this as part of clearing up a bit of the mess that generics have brought
with them.