First attempt at better defaulting (#70). Lists seem to work.
To begin with, the emitter now has the idea of a 'floating' adjustment for
generics. Signatures from emitter's sig_stack_pos to the adjustment are
considered in use by the current generic call. This is an important and
necessary change, because it allows eval_build_list to look back to determine
what a generic should be.

Additionally, when a function is entered, it reserves that many generics so
that eval_build_list can access them right.

Consider this:
* list[integer] l = []

* list[list[integer]] l = []    l.append([])
  A is resolved as list[integer]. An empty list of integer is made.

* function k[A](list[A] l, A val) { l.append(val) }
  concretely known.

The generic adjust is a bit ugly, but there needs to be some way for list
evaluation to be able to access what a generic should really be.

The good parts:
* It seems that the 'try to fix it back up' part of type matching can go away
  as part of this. I've already removed the list case, with no issues.
* It is now possible to do 'list[integer] l = []'. This will be vital later on
  when declarations mandate assignments.

What's left:
* It needs to work for hashes and tuples.
* There's an opportunity to make signature building in emitter a bit better
  and simpler (for most cases, that is).
* Documentation needs to mention that [] is a list of any when there is no
  type information to use.