The lexer can now enter multiple sources.
Previously, the lexer allowed trying to enter a file/string/other after
something was already entered. However, doing so would result in a crash.

Unfortunately, this isn't an import mechanism, but it's a critical part OF an
import mechanism (so that's a plus). I did this in part because it's needed
for #68.

Additionally, a decision has been made that will be important in the future:
The lexer will only honor the mode of the first file entered, and enforce a
no-tags policy on entered files. For now, this is to allow the parser to load
tagless things (when I to #68) without having to save/restore the mode.

For now, this is just about adding a mechanism which will do nifty things in
the future, but nothing right now.