Lexer no longer uses hit_eof to calculate line bumps.
If an eof has been seen, the reading functions now check if the eof is at the
start of a line or not. If it isn't the start of a line (meaning that code
terminates with an eof instead of a newline), then bump the line number as if
it were a newline. Otherwise, don't bump the line number.

The lexer's hit_eof was causing problems with loading/unloading files since
a loaded file will always have an eof at the end. This would causes the
original file to never increment the line number for the last line if the eof
was not at the end.

Cutting away hit_eof is just a bonus.