Enum classes (and any) can now directly compare to subtypes.
Consider this:
    any someval = 10

It is now possible to do this:
    someval == 10
    or
    10 == someval

The same is also now possible with enum classes. An Option class that SomeType
and NoneType members can be compared directly with a SomeType or NoneType
instance.

On the downside, a temporary hack has been put in place to keep the vm from
crashing when garbage collecting enum class instances. This will be fixed in
time (the real solution is to properly look at classes and signatures to
determine if they can be circular).

A couple tests have been added to test/pass/sanity.ly as well.