Remove automatic conversion between integer and double.
Instead, add double::to_integer to complement a new integer::to_double. This
was done because the integer<->double conversion could prove strange when
combined with an enum class.

Additionally, this feature is not, and never was recursable (a list of
integers would not be auto-converted to a list of doubles, for example). This
felt like more of a hack over time, and additionally felt rather useless.

Test removed from test/pass/sanity.ly.