o_assign -> o_fast_assign. o_{ref,any}_assign combined into o_box_assign.
This reflects the fact that the regular o_assign is very rare (only for
integer and double), and also simpler than any other kind of assign.

o_box_assign now handles simple ref/deref situations, as well as enum class
assignments and any assignments. The reason these were combined is that ref
assign is not a very common opcode (it will only trigger when assigning one
var to another var). It's not common because the interpreter folds down a LOT
of stuff to avoid having to write assignments.

This drops an opcode, simplifying the interpreter a bit, which is nice.