How about another try at enum classes?
So...it turns out that I had some misconceptions about how enum classes and
variants should work. There's no functionality change, though things seem less
likely to crash this time.

Enum classes now create classes instead of needing classes. Additionally,
there is no need to use ::new to create instances of a variant. This means
that usage of enum classes and variants is simpler. In going with the Option
class example, it's now possible to use Some(x) instead of SomeType::new(x).
More importantly, 'None' can be used in place of NoneType::new().

Unfortunately, variants kind of break some previous Lily things. For one,
trying to 'show' a variant that takes args (even though it looks like a
function) will not show function code since a function is not used to create
variant values.

Variants cannot be used with the wrong number of args either (so the use of
None() would be invalid). I don't like the non-showability of variants, but
not enough to roll this back.

Variant-related tests updated too.

Warning: Variant-related code is still fragile. Be careful.