Massive change! Variables now demand an initializing assignment.
Now that enum classes have landed, it's now possible to use an Option with
Some and None to represent variables which may or may not be initialized.

Because of this, Lily now DEMANDS that ALL variables have an initializing
assignment. I do not expect this to change, since I have been really wanting
this for a long time now.

All tests have been updated to initialize variables on assignment. The emitter
now disables a variable being used to initialize itself too. I deleted one
test because it involved nil-ness (which is a now-dying concept in the
interpreter).