Only primitive values can be condition cases now.
Previously, a condition using an 'any' would work. It doesn't now. Conditions
must be either an integer, a double, a string, or a list.

Lists are considered truthy based on a count of elements. Strings are
considered truthy based on their size. For that case, the literal optimization
was fixed to reflect that (previously, an empty string was truthy).

o_jump_if no longer checks for nil-ity or to unbox 'any' values. This is
another step toward erasing nil-ity.