The assumption that user vars can be nil is gone.
This assumption has been baked into the code from the very beginning. Combined
with the other changes today, it marks a massive change in how Lily works
that has been ingrained in Lily since it was made.

Intermediate values can still be nil (and this is the reason that there is
still a fair amount of nil-checking code that will never die). However, the
user should never see any values that are nil.

With this commit, I wave goodbye to user-level nil-ity.