It begins! Extract passing tests from pre-commit, touchup blastmaster.py
pre-commit-hook.py now contains ONLY tests that fail. It will soon be
completely purged.

blastmaster.py now checks for output and considers any output to be a failure.
This is a temporary hack (much like how it comments out the sanity test).

One big change has been made (for now): the main runner (lily aka lily_main.c)
now runs files in no-tag mode. I'll eventually include an option to allow
tags if the user wants that (partly because some tests would need tags to be
on to be testable).