sanity.ly is dead.
Most of the passing tests from it have been turned into smaller tests and
thrown into test/pass somewhere.

The reason for this change is that it is easier to isolate what exactly failed
when there are a lot of varying small tests. One of the constant problems I
had with the sanity.ly test is that something would break internally and I'd
have no idea where to start cutting to reveal the problem.