Emitter: template check now does partial match for enum by a variant.
Suppose we've got a setup like this:
    enum class Option[A] {
        Some(A),
        None
    }

    function f[A](Option[A] first, Option[A] second) { }

It would be nice to do this:

    f(Some(10), None)
    f(None, None)
    f(None, Some(10))

The emitter now allows this (but does not box the values, so this is sorta
broken). The idea is that if there's a variant that doesn't have generics to
offer, then it's allowed and no generics are defined by it being used.

In the first case, Some(10) results in the A of Option[A] being defined as
integer. The None in the first case is allowed because it does not conflict
with the current understanding of Option containing an integer.

In the second case, A is never defined, and (when this is all fixed), will
be transformed into any. The function will see Option[any].

In the third case, None does not specify any generics, but Some(10) does. Both
will be passed as Option[integer].

This patch DOES restrict improper definitions, however:

    f(Some(10), Some("10"))

is an error because A is defined as integer, then attempted to be defined as
a string. Unlike Scala, Lily isn't smart enough to make A into any in this
case. I'm not sure I want that behavior, either.