The interpreter can now handle variants and generics.
Consider this setup:
    enum class Option[A] { Some(A), None }
    function f[A](Option[A] first, Option[A] second) { }

In such a case, it would be nice to do this:

    f(None, None) # A = any
    f(Some(10), Some(10)) # A = integer
    f(Some("10"), None) # A = string

With this commit, the following now properly works and reboxes the variants
into proper enum class values. I've added two tests: One that checks for this
with regular calling, and one that tests using varargs. Both work.

Part of this commit was adjusting the symtab's finalize_sig to now mark all
function signatures which contain an enum class as one of their parameters to
make the emitter's job of checking things easier.