The emitter can now handle more complex enum classes.
I finally went back and fixed the check for a proper enum variant to do what
it really ought to do (the same thing that template_check does: Check that
the variants that are there match up, use type 'any' for the rest).

With this commit, it's now possible to do things like 'var v = None' and have
it automatically worked out to 'Option[any]'.

A bug was fixed in symtab's variant result calculation: It was writing the
generics down to position 0. This has been fixed so it calculates the proper
variant result.

Addded a new test with a more complex enum class, and updating an existing
test to use 'var' with 'None' to test this new thing.