The interpreter can now handle static lists+hashes with variants.
Previously, the interpreter would allow bare variants in static lists and
hashes. This is a problem, because the interpreter doesn't like bare variants
and it isn't consistent with previous behavior.

In the event a static list/(or the values of a static hash) have variants or
enums, the interpreter will call a function to fix them to the same type.

If all variants+enum class values agree on a type, then that becomes the type
of the values.

If it is a collection of variants and generics of the enum class are missing,
then the missing generics are filled with type 'any'. This can happen when,
say, with an Option enum class where a static list contains [None, None, None]
(which would become list[Option[any]] now).

If the variants and enums do not agree on a type, then each variant is put
into an enum based on what information it has alone. Missing generic
information is filled with 'any', as usual. All values are then put into an
'any'.

Additionally, emit_rebox_value now checks if the result is an 'any' and the
input value is a variant. If the input is a variant, the variant is first put
into an enum class value (with missing generics filled in with type 'any')
before being put into an 'any'. This prevents bare variants from being visible
within an 'any' (which is inconsistent with other parts of the interpreter).

Symtab: A bug was fixed in calculating the siglist_size of a 'constructor'
return signature. This was fixed because it was crashing the above code. All
tests show that this did not break anything.

A test has been added to make sure variants are defaulting correctly within
a static list (and a static hash).