Bugfix: `hash[integer, any] h = [1 => 1] failed`.
The problem was embarassingly simple: hash building was checking if hash
values were mixed (and making them into any's if that was the case). However,
it was not checking if the value was expected to be 'any' (and making values
into anys in that case).

Test added, because why not?