Match has landed. This finally closes #67.
This commit introduces support for a match operation that works like a
conditional block (match <expr>: { ... }).

All match blocks must always be multi-line blocks. I decided to do this
because I have made all 'case' conditions within match automatically
multi-line so that there won't be a pollution of { and } between cases in a
match expression.

(Assuming 'enum class Option[A] { Some(A), None }'

What works:
* Decomposition. A 'Some' value will require one identifier since it has one
  value contained with it. Lily, in following with other languages that offer
  this match feature, does not require that types be specified. It does type
  inference for the variables.
* Tests have been added to make sure that match cannot try to decompose a
  variant without a value, and to not decompose something that should be
  decomposed.
* If the match block is incomplete, a message is printed saying that it is,
  as well as what variants have not been cased for.
* Values at the end of each match case scope out. So a 'None' case won't see
  what the Some decomposed to, or any values declared within the Some case.
* Match cases do not have to specify code, and is intentional. My reasoning
  behind this is that a case without code is a way of saying that "I don't
  need to do anything here." However, a missing case is assumed to be a
  mistake in forgetting to add a case after having added another variant to
  an enum.

There is no support for the use of, say, _ to indicate that a decomposition
value should be ignored. I intend to add that at some point in the future.