Emitter: eval_call should not check for ast->result == NULL.
A long time ago, the parser used to handle calls of something that was known
to be a function by making tree_call and setting the tree's result to that
particular function's var.

This was changed when the parser's expression handling was written, so that
calling known function calls would no longer be a special case.

In short, it was ALWAYS triggered and therefore useless.